# swagger.yaml
consumes:
    - application/json
produces:
    - application/json
schemes:
    - https
swagger: "2.0"
info:
    contact:
        email: test@trahan.dev
        name: Justin Trahan
    license:
        name: N/A
    termsOfService: there are no TOS at this moment, use at your own risk we take no responsibility
    title: go-infra API.
    version: v1.1.0
host: infra.trahan.dev
paths:
...
    /external-applications:
        get:
            operationId: getAllExternalApplications
            responses:
                "200":
                    $ref: '#/responses/GetAllExternalApplicationsResponse'
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Get all external applications.
            tags:
                - external-applications
        post:
            operationId: createExternalApplication
            parameters:
                - in: body
                  name: body
                  schema:
                    $ref: '#/definitions/CreateExternalApplicationRequest'
                  x-go-name: Body
            responses:
                "201":
                    $ref: '#/responses/CreateExternalApplicationResponse'
                "400":
                    description: Bad request - invalid input data
                "409":
                    description: Conflict - application with this name already exists
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Create a new external application.
            tags:
                - external-applications
    /external-applications/by-name/{name}:
        delete:
            operationId: deleteExternalApplicationByName
            parameters:
                - description: Name of the external application
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
            responses:
                "204":
                    description: External application deleted successfully
                "404":
                    description: External application not found
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Delete an external application by name.
            tags:
                - external-applications
        get:
            operationId: getExternalApplicationByName
            parameters:
                - description: Name of the external application
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
            responses:
                "200":
                    $ref: '#/responses/GetExternalApplicationByNameResponse'
                "404":
                    description: External application not found
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Get an external application by name.
            tags:
                - external-applications
    /external-applications/id/{name}:
        get:
            operationId: getExternalApplicationIdByName
            parameters:
                - description: Name of the external application
                  in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
            responses:
                "200":
                    $ref: '#/responses/GetExternalApplicationIdByNameResponse'
                "404":
                    description: External application not found
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Get an external application ID by name.
            tags:
                - external-applications
    /external-applications/name/{ID}:
        get:
            operationId: getExternalApplicationNameById
            parameters:
                - description: ID of the external application
                  in: path
                  name: ID
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/GetExternalApplicationNameByIdResponse'
                "400":
                    description: Bad request - invalid UUID format
                "404":
                    description: External application not found
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Get an external application name by ID.
            tags:
                - external-applications
    /external-applications/{ID}:
        delete:
            operationId: deleteExternalApplicationById
            parameters:
                - description: ID of the external application
                  in: path
                  name: ID
                  required: true
                  type: string
            responses:
                "204":
                    description: External application deleted successfully
                "400":
                    description: Bad request - invalid UUID format
                "404":
                    description: External application not found
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Delete an external application by ID.
            tags:
                - external-applications
        get:
            operationId: getExternalApplicationById
            parameters:
                - description: ID of the external application
                  in: path
                  name: ID
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/GetExternalApplicationByIdResponse'
                "400":
                    description: Bad request - invalid UUID format
                "404":
                    description: External application not found
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Get an external application by ID.
            tags:
                - external-applications
        put:
            operationId: updateExternalApplication
            parameters:
                - description: ID of the external application
                  in: path
                  name: ID
                  required: true
                  type: string
                - in: body
                  name: body
                  schema:
                    $ref: '#/definitions/UpdateExternalApplicationRequest'
                  x-go-name: Body
            responses:
                "200":
                    $ref: '#/responses/UpdateExternalApplicationResponse'
                "400":
                    description: Bad request - invalid UUID format or request body
                "404":
                    description: External application not found
                "500":
                    description: Internal server error
            security:
                - bearer: []
            summary: Update an external application.
            tags:
                - external-applications
...
definitions:
...
    CreateExternalApplicationRequest:
        properties:
            appDescription:
                type: string
                x-go-name: AppDescription
            endpointUrl:
                type: string
                x-go-name: EndpointUrl
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/babbage88/go-infra/services/external_applications
...
    ExternalApplicationDao:
        properties:
            appDescription:
                type: string
                x-go-name: AppDescription
            createdAt:
                format: date-time
                type: string
                x-go-name: CreatedAt
            endpointUrl:
                type: string
                x-go-name: EndpointUrl
            id:
                format: uuid
                type: string
                x-go-name: Id
            lastModified:
                format: date-time
                type: string
                x-go-name: LastModified
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/babbage88/go-infra/services/external_applications
...
    UpdateExternalApplicationRequest:
        properties:
            appDescription:
                type: string
                x-go-name: AppDescription
            endpointUrl:
                type: string
                x-go-name: EndpointUrl
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/babbage88/go-infra/services/external_applications
...
responses:
...
    CreateExternalApplicationResponse:
        description: (empty)
        schema:
            $ref: '#/definitions/ExternalApplicationDao'
...
    GetAllExternalApplicationsResponse:
        description: (empty)
        schema:
            items:
                $ref: '#/definitions/ExternalApplicationDao'
            type: array
...
    GetExternalApplicationByIdResponse:
        description: (empty)
        schema:
            $ref: '#/definitions/ExternalApplicationDao'
    GetExternalApplicationByNameResponse:
        description: (empty)
        schema:
            $ref: '#/definitions/ExternalApplicationDao'
    GetExternalApplicationIdByNameResponse:
        description: (empty)
        schema:
            properties:
                id:
                    format: uuid
                    type: string
                    x-go-name: ID
            type: object
    GetExternalApplicationNameByIdResponse:
        description: (empty)
        schema:
            properties:
                name:
                    type: string
                    x-go-name: Name
            type: object
...
    UpdateExternalApplicationResponse:
        description: (empty)
        schema:
            $ref: '#/definitions/ExternalApplicationDao'
...
securityDefinitions:
    bearer:
        in: header
        name: Authorization
        type: apiKey
security:
    - bearer: []